feat: move lucia from auth to server directory
