+++ apps/client/src/hooks.server.ts
@@ -1,4 +1,4 @@
-import { auth } from '$lib/auth/lucia';
+import { auth } from '$lib/server/lucia';
 import type { Handle } from '@sveltejs/kit';
 
 export const handle: Handle = async ({ event, resolve }) => {
